# Fix the const correctness of the DeviceTensor class.