--- a/general/forall.hpp
+++ b/general/forall.hpp
@@ -51,21 +51,21 @@ const int MAX_Q1D = 14;
 // The MFEM_FORALL wrapper
 #define MFEM_FORALL(i,N,...)                             \
    ForallWrap<1>(true,N,                                 \
-                 [=] MFEM_DEVICE (int i) {__VA_ARGS__},  \
-                 [&]             (int i) {__VA_ARGS__})
+                 [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
+                 [&]             (int i) mutable {__VA_ARGS__})
 
 // MFEM_FORALL with a 2D CUDA block
 #define MFEM_FORALL_2D(i,N,X,Y,BZ,...)                   \
    ForallWrap<2>(true,N,                                 \
-                 [=] MFEM_DEVICE (int i) {__VA_ARGS__},  \
-                 [&]             (int i) {__VA_ARGS__},  \
+                 [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
+                 [&]             (int i) mutable {__VA_ARGS__},  \
                  X,Y,BZ)
 
 // MFEM_FORALL with a 3D CUDA block
 #define MFEM_FORALL_3D(i,N,X,Y,Z,...)                    \
    ForallWrap<3>(true,N,                                 \
-                 [=] MFEM_DEVICE (int i) {__VA_ARGS__},  \
-                 [&]             (int i) {__VA_ARGS__},  \
+                 [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
+                 [&]             (int i) mutable {__VA_ARGS__},  \
                  X,Y,Z)
 
 // MFEM_FORALL that uses the basic CPU backend when use_dev is false. See for
@@ -73,8 +73,8 @@ const int MAX_Q1D = 14;
 // device for operations on small vectors.
 #define MFEM_FORALL_SWITCH(use_dev,i,N,...)              \
    ForallWrap<1>(use_dev,N,                              \
-                 [=] MFEM_DEVICE (int i) {__VA_ARGS__},  \
-                 [&]             (int i) {__VA_ARGS__})
+                 [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
+                 [&]             (int i) mutable {__VA_ARGS__})
 
 
 /// OpenMP backend