--- a/general/forall.hpp
+++ b/general/forall.hpp
@@ -43,20 +43,20 @@ const int MAX_Q1D = 14;
 #define MFEM_FORALL(i,N,...)                             \
    ForallWrap<1>(true,N,                                 \
                  [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
-                 [&] MFEM_LAMBDA (int i) mutable {__VA_ARGS__})
+                 [&] MFEM_LAMBDA (int i) {__VA_ARGS__})
 
 // MFEM_FORALL with a 2D CUDA block
 #define MFEM_FORALL_2D(i,N,X,Y,BZ,...)                   \
    ForallWrap<2>(true,N,                                 \
                  [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
-                 [&] MFEM_LAMBDA (int i) mutable {__VA_ARGS__},\
+                 [&] MFEM_LAMBDA (int i) {__VA_ARGS__},\
                  X,Y,BZ)
 
 // MFEM_FORALL with a 3D CUDA block
 #define MFEM_FORALL_3D(i,N,X,Y,Z,...)                    \
    ForallWrap<3>(true,N,                                 \
                  [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
-                 [&] MFEM_LAMBDA (int i) mutable {__VA_ARGS__},\
+                 [&] MFEM_LAMBDA (int i) {__VA_ARGS__},\
                  X,Y,Z)
 
 // MFEM_FORALL that uses the basic CPU backend when use_dev is false. See for
@@ -65,7 +65,7 @@ const int MAX_Q1D = 14;
 #define MFEM_FORALL_SWITCH(use_dev,i,N,...)              \
    ForallWrap<1>(use_dev,N,                              \
                  [=] MFEM_DEVICE (int i) mutable {__VA_ARGS__},  \
-                 [&] MFEM_LAMBDA (int i) mutable {__VA_ARGS__})
+                 [&] MFEM_LAMBDA (int i) {__VA_ARGS__})
 
 
 /// OpenMP backend