--- a/linalg/dtensor.hpp
+++ b/linalg/dtensor.hpp
@@ -114,14 +114,28 @@ public:
 
    /// Const accessor for the data
    template <typename... Args> MFEM_HOST_DEVICE inline
-   Scalar& operator()(Args... args) const
+   const Scalar& operator()(Args... args) const
+   {
+      static_assert(sizeof...(args) == Dim, "Wrong number of arguments");
+      return data[ TensorInd<1, Dim, Args...>::result(sizes, args...) ];
+   }
+
+   /// Accessor for the data
+   template <typename... Args> MFEM_HOST_DEVICE inline
+   Scalar& operator()(Args... args)
    {
       static_assert(sizeof...(args) == Dim, "Wrong number of arguments");
       return data[ TensorInd<1, Dim, Args...>::result(sizes, args...) ];
    }
 
    /// Subscript operator where the tensor is viewed as a 1D array.
-   MFEM_HOST_DEVICE inline Scalar& operator[](int i) const
+   MFEM_HOST_DEVICE inline const Scalar& operator[](int i) const
+   {
+      return data[i];
+   }
+
+   /// Subscript operator where the tensor is viewed as a 1D array.
+   MFEM_HOST_DEVICE inline Scalar& operator[](int i)
    {
       return data[i];
    }
@@ -136,6 +150,11 @@ inline DeviceTensor<sizeof...(Dims),T> Reshape(T *ptr, Dims... dims)
    return DeviceTensor<sizeof...(Dims),T>(ptr, dims...);
 }
 
+template <typename T, typename... Dims>
+inline const DeviceTensor<sizeof...(Dims),T> Reshape(const T *ptr, Dims... dims)
+{
+   return DeviceTensor<sizeof...(Dims),T>(const_cast<T*>(ptr), dims...);
+}
 
 typedef DeviceTensor<1,int> DeviceArray;
 typedef DeviceTensor<1,double> DeviceVector;