The `DeviceTensor` code subverts the `const` correctness which can result in very difficult to find bugs.
<!--GHEX{"id":1626,"author":"YohannDudouit","editor":"tzanio","reviewers":["jakubcerveny","v-dobrev"],"assignment":"2020-07-21T11:41:20-07:00","approval":"2020-09-01T21:22:17.902Z","merge":"2020-09-08T21:22:17.902Z"}XEHG--><!--GHEXTABLE-->
 | PR | Author | Editor | Reviewers | Assignment | Approval | Merge| 
 | --- | --- | --- | --- | --- | --- | --- | 
| [#1626](https://github.com/mfem/mfem/pull/1626) | @YohannDudouit | @tzanio | @jakubcerveny + @v-dobrev | 07/21/20 | 09/01/20 | ⌛due 09/08/20 | |
<!--ELBATXEHG-->