@v-dobrev 2020-07-16
The current definition of `DeviceTensor` does not violate const-correctness for the data: if you construct a `DeviceTensor<const double>` object from `const double *` you cannot change its entries -- the subscript operators (both `()` and `[]`) return a reference to `const double`. What you can change potentially is the `DeviceTensor` object itself (e.g. to make it use another `const double *`), however, the class does not have any such methods at the moment. In this regard, the class is similar to a pointer -- the constness of the pointer is separate from the constness of the type it points to. I think this type of definition is quite suitable here since `DeviceTensor` serves as a view for externally managed data.

Also, note that you cannot construct `DeviceTensor<double>` from `const double *` unless you explicitly cast the pointer as `double *`. On the other hand, it is okay to construct `DeviceTensor<const double>` from `double *` as expected.