@YohannDudouit 2020-07-20
@v-dobrev The fact that it's correct in one case does not mean it's correct in general.
The accessor has the following signature: `Scalar& operator(int... args) const`, this is returning a non-`const` reference and the method claims that it is `const` for the class, this is a very dangerous thing to do. It means that we can modify `const DeviceTensor`, and the compiler will think that we did not. This can also result in almost impossible to track bugs, because when you allow to modify `const` objects you can't trust `const` any more.