@v-dobrev 2020-07-21
> It means that we can modify `const DeviceTensor`, and the compiler will think that we did not.

The compiler cannot (and I'm sure it will not) make such assumptions -- the contents of the pointer is not covered by the constness of the object -- just like with `double * const` pointer.

That said, I agree that allowing the data of an object `const DeviceTensor<double>` to be modified is not standard and takes a little getting used to -- you just need to remember to think of it as a `double * const` pointer replacement that supports multi-indexing.

In some regards, `DeviceTensor` (in its current form) is similar to the c++20 `std::span` class template (https://en.cppreference.com/w/cpp/container/span) where the indexing operator has similar signature -- it is a const method that returns a reference to the template type (https://en.cppreference.com/w/cpp/container/span/operator_at). Another similarity is that both classes do not own their data.

In the end, this is a design preference, so if others also prefer the change you propose, I'll be fine with it.