@YohannDudouit 2020-07-21
I'm very confused by what you are saying, if we have a class `A`,  almost _everyone_ expect `const A a` to mean that you cannot change the state of `a`.