@v-dobrev 2020-08-07
As I commented before, I don't see a good reason for this change -- the protection of the data is provided by the `const` qualifier in the `Scalar` template type. Additionally, the const-ness in the proposed modification is easy to break, due to the shallow copy used in the copy-constructor.

At this point, I personally prefer not to merge this. However, in a way, this PR already has two approvals (@YohannDudouit and @jakubcerveny), so my preference is overruled, unless @tzanio agrees with me. If that's the case then we'll need a 5th person to break the tie.