@YohannDudouit 2020-08-07
I think having bugs being caught by the compiler instead of having to find them with a debugger is a pretty strong argument.