@YohannDudouit 2020-08-08
@camiers here are the performance comparison on `ex1 -pa -d cuda -m ../data/fichera.mesh -o 4`:
On master:
```
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   46.51%  338.82ms       402  842.85us  757.82us  1.3410ms  _ZN4mfem10CuKernel3DIZNS_22SmemPADiffusionApply3DILi5ELi6EEEviRKNS_5ArrayIdEES5_RKNS_6VectorES8_RS6_iiEUliE_EEviT_
                    9.55%  69.604ms      1199  58.051us  56.799us  60.991us  _ZN4mfem10CuKernel1DIZNS_3addERKNS_6VectorEdS3_RS1_EUliE_EEviT_
                    8.94%  65.099ms       402  161.94us  160.10us  164.93us  _ZN4mfem10CuKernel1DIZNKS_18ElementRestriction13MultTransposeERKNS_6VectorERS2_EUliE_EEviT_
                    5.58%  40.640ms         1  40.640ms  40.640ms  40.640ms  _ZN4mfem10CuKernel1DIZNS_22QuadratureInterpolator6Eval3DILi3ELi0ELi0EEEviiRKNS_9DofToQuadERKNS_6VectorERS6_S9_S9_iEUliE_EEviT_
```
On this branch:
```
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   46.73%  339.43ms       402  844.35us  759.61us  1.3685ms  _ZN4mfem10CuKernel3DIZNS_22SmemPADiffusionApply3DILi5ELi6EEEviRKNS_5ArrayIdEES5_RKNS_6VectorES8_RS6_iiEUliE_EEviT_
                    9.57%  69.496ms      1199  57.961us  56.928us  59.936us  _ZN4mfem10CuKernel1DIZNS_3addERKNS_6VectorEdS3_RS1_EUliE_EEviT_
                    8.98%  65.260ms       402  162.34us  159.97us  166.50us  _ZN4mfem10CuKernel1DIZNKS_18ElementRestriction13MultTransposeERKNS_6VectorERS2_EUliE_EEviT_
                    5.19%  37.707ms         1  37.707ms  37.707ms  37.707ms  _ZN4mfem10CuKernel1DIZNS_22QuadratureInterpolator6Eval3DILi3ELi0ELi0EEEviiRKNS_9DofToQuadERKNS_6VectorERS6_S9_S9_iEUliE_EEviT_
```