@YohannDudouit 2020-08-24
@v-dobrev can you please answer these questions:
- Why do you consider that modifying the values inside a `const DeviceTensor` does not modify the `DeviceTensor`?
- Why is catching bug at compilation instead of runtime not an argument to you?
- Is there any advantage of the current design over the proposed one?
- Is there any disadvantage of the proposed design over the current one?