@tzanio 2020-09-02
There are a number of failures in tonight's autotest runs on `tux429`:

```
linalg/../general/forall.hpp(121): error: no instance of function "lambda [](int) mutable->void" matches the argument list and object (the object has type qualifiers that prevent a match)
            argument types are: (const int)
            object type is: const lambda [](int) mutable->void
          detected during:
            instantiation of "void mfem::RajaCudaWrap2D(int, DBODY &&, int, int, int) [with DBODY=lambda [](int) mutable->void &]" 
(316): here
            instantiation of "void mfem::ForallWrap<DIM,DBODY,HBODY>(__nv_bool, int, DBODY &&, HBODY &&, int, int, int) [with DIM=1, DBODY=lambda [](int) mutable->void, HBODY=lambda [](int)->void]" 
linalg/operator.cpp(429): here

linalg/../general/forall.hpp(140): error: no instance of function "lambda [](int) mutable->void" matches the argument list and object (the object has type qualifiers that prevent a match)
            argument types are: (const int)
            object type is: const lambda [](int) mutable->void
          detected during:
            instantiation of "void mfem::RajaCudaWrap3D(int, DBODY &&, int, int, int) [with DBODY=lambda [](int) mutable->void &]" 
(319): here
            instantiation of "void mfem::ForallWrap<DIM,DBODY,HBODY>(__nv_bool, int, DBODY &&, HBODY &&, int, int, int) [with DIM=1, DBODY=lambda [](int) mutable->void, HBODY=lambda [](int)->void]" 
linalg/operator.cpp(429): here
```