@v-dobrev 2020-09-03
Re-merged in `next`.